Switch to new InventoryManager
- remove InventoryManager
- rename InventoryMAnagerV2 to InventoryManager
- update usings
- fix ProductBuilderTests build fail
