+++ src/Foundation/Connect/code/Managers/Inventory/InventoryManager.cs
@@ -1,4 +1,4 @@
-//    Copyright 2020 EPAM Systems, Inc.
+﻿//    Copyright 2020 EPAM Systems, Inc.
 // 
 //    Licensed under the Apache License, Version 2.0 (the "License");
 //    you may not use this file except in compliance with the License.
@@ -12,9 +12,12 @@
 //    See the License for the specific language governing permissions and
 //    limitations under the License.
 
-namespace Wooli.Foundation.Connect.Managers
+namespace Wooli.Foundation.Connect.Managers.Inventory
 {
     using System.Collections.Generic;
+
+    using Base.Models.Logging;
+    using Base.Services.Logging;
 
     using DependencyInjection;
 
@@ -25,32 +28,31 @@ namespace Wooli.Foundation.Connect.Managers
     using Sitecore.Commerce.Services.Inventory;
     using Sitecore.Diagnostics;
 
-    [Service(typeof(IInventoryManager))]
-    public class InventoryManager : IInventoryManager
+    [Service(typeof(IInventoryManager), Lifetime = Lifetime.Singleton)]
+    public class InventoryManager : BaseManager, IInventoryManager
     {
         private readonly InventoryServiceProvider inventoryServiceProvider;
 
-        public InventoryManager(IConnectServiceProvider connectServiceProvider)
+        public InventoryManager(IConnectServiceProvider connectServiceProvider, ILogService<CommonLog> logService)
+            : base(logService)
         {
             Assert.ArgumentNotNull(connectServiceProvider, nameof(connectServiceProvider));
+
             this.inventoryServiceProvider = connectServiceProvider.GetInventoryServiceProvider();
         }
 
-        public ManagerResponse<GetStockInformationResult, IEnumerable<StockInformation>> GetStockInformation(
+        public GetStockInformationResult GetStockInformation(
             string shopName,
             IEnumerable<CommerceInventoryProduct> inventoryProducts,
             StockDetailsLevel detailsLevel)
         {
+            Assert.ArgumentNotNullOrEmpty(shopName, nameof(shopName));
             Assert.ArgumentNotNull(inventoryProducts, nameof(inventoryProducts));
+            Assert.ArgumentNotNull(detailsLevel, nameof(detailsLevel));
 
             var request = new GetStockInformationRequest(shopName, inventoryProducts, detailsLevel);
-            request.Location = string.Empty;
 
-            var stockInformation = this.inventoryServiceProvider.GetStockInformation(request);
-
-            return new ManagerResponse<GetStockInformationResult, IEnumerable<StockInformation>>(
-                stockInformation,
-                stockInformation.StockInformation ?? new List<StockInformation>());
+            return this.Execute(request, this.inventoryServiceProvider.GetStockInformation);
         }
     }
 }
