+++ src/Foundation/Connect/tests/Wooli.Foundation.Connect.Tests.csproj
@@ -564,7 +564,7 @@
     <Compile Include="Managers\Account\AccountManagerV2Tests.cs" />
     <Compile Include="Managers\BaseManagerTests.cs" />
     <Compile Include="Managers\Cart\CartManagerV2Tests.cs" />
-    <Compile Include="Managers\Inventory\InventoryManagerV2Tests.cs" />
+    <Compile Include="Managers\Inventory\InventoryManagerTests.cs" />
     <Compile Include="Managers\Order\OrderManagerV2Tests.cs" />
     <Compile Include="Managers\Payment\PaymentManagerV2Tests.cs" />
     <Compile Include="Mappers\Profiles\PaymentProfileTests.cs" />
